requirements.txt

pandas
sqlalchemy
sqlite
dbt
sqlite3
smtplib
email
apache-airflow
apache-airflow-dbt